rename parser, try to fix gcc error
